+++ src/lib/notifications/api/NotificationsApi.web.js
@@ -1,3 +1,5 @@
 export const NotificationsAPI = {
   isSupported: () => false,
+  // eslint-disable-next-line require-await
+  getInitialNotification: async () => null,
 }
